Don't use system() to call the process from itself anymore.
I had thought this to be necessary originally when I saw that the process re-launching itself as a child process using system() evaded crashing from reading inaccessible memory.  Reading the documentation on system()'s implementation on Linux, I have come to find that in this implementation system() already disables the relevant signals for us, which is something we can do in standard C code without relying on something as monstrously inefficient as system().